refactor(ui): extract ChatComposer from ChatView
After the sidebar carve-out, the next-biggest concrete cut is the
composer's <form> subtree plus its handler graph plus the three
inline notice components that only the composer renders. They go
together: handleSubmit reads textareaRef, handleKeyDown reads
formRef, handleMessageHistoryKey reads messageHistory and the two
local history refs, and the three notices (ChatErrorPanel,
ChatSetupRepairNotice, SelectedModelReadinessNotice) sit at the
top of the form before the textarea.

What moves into ChatComposer.tsx:
  - The 285-line <form> JSX (notices, queued-message editor,
    message-controls bar, textarea + send/stop button, "agent
    busy" footer, mod-enter toggle).
  - Composer-local state: modEnterMode (persisted), formRef,
    messageHistoryCursorRef, messageHistoryDraftRef.
  - Handlers: setComposerText, handleMessageChange,
    handleMessageHistoryKey, handleKeyDown, handleSubmit,
    toggleModEnterMode.
  - History-reset effect (resets the two history refs on
    activeSessionID change — scroll-side reset stays in ChatView
    since it concerns the transcript surface).
  - Notice components: ChatErrorPanel, ChatSetupRepairNotice,
    SelectedModelReadinessNotice (with InfoChip,
    selectedModelNoticeDetails, repairActionIcon).
  - Helpers: compactID (the error panel uses it for request/trace
    IDs), providerLabelForHecateChat (only used in the composer's
    message-controls bar).

What stays in ChatView:
  - All the derived-value plumbing (composerVisible,
    composerRepair, sendDisabled, hecateChatModelReady,
    selectableModels, hecateProviderOptions, …) — header / modals
    / transcript / empty-state read most of them too, so a clean
    move would need a shared derived-state hook. Keep them on
    the ChatView side and pass the bag down for now.
  - Cross-region coordination: textareaRef (created here so
    ChatSidebar's onSelectSession can focus it), focus effects,
    scroll state.
  - Composer-form callbacks ChatView owns: chooseWorkspace,
    enableToolsForSelectedModel, openClaudeCodeSetup.

Imports retire from ChatView: SyntheticEvent, formatErrorCode,
usePersistedState, ChatNoticeInline, ClaudeCodePreflightCard,
claudeCodeSetupTokenCommand, ExternalAgentConfigControls,
LockedHecateModelSnapshot, HecateProviderConfigControl,
HecateModelConfigControl, RuntimeConsoleViewModel type.

ChatView re-imports from ChatComposer: ChatComposer,
SelectedModelReadinessNotice (ChatEmptyState's compact variant),
compactID (HecateTaskApprovalsBanner + formatTaskLinkLabel),
repairActionIcon (ChatEmptyState's repair CTA).

Diff:
  - ChatView.tsx: 3130 → 2526 LOC (-604).
  - ChatComposer.tsx: new, 783 LOC.
  - Combined: +179 LOC. The Props bag is the cost — extracting it
    cleanly later wants a shared derived-state hook.

Verification:
  - bun run typecheck — clean (tsgo -b).
  - bunx vitest run — 824/824 pass (38 files, no count delta).
  - bun run test:e2e — 60/60 chromium pass.
  - Manual: dev server in Tauri webview, sidebar + topbar
    unchanged; clicking "New Hecate chat" opens the draft and
    the empty-state "Choose a workspace" repair shows (composer
    correctly hidden because no providers are configured). No
    console errors.
